Make `test` work for 1.0.0-M1 and 0.6.19.
It does not work for Scala.js 0.6.20 because the test interface
protocol was changed in 0.6.20. Since we always use the sbt test
adapter of Scala.js 1.0.0-M1, it can only communicate to the JS
side if it uses a compatible protocol.

When Scala.js 1.0.0-M2 is released with the new protocol that is
used in 0.6.20, we will be able to upgrade to supporting `test`
for those new versions, but at the cost of breaking it for the
older versions.