Upgrade to Scala.js 1.0.0-M8 and 0.6.28.
Support for older versions is dropped.

For Scala.js 0.6.x, we now inject our own test bridge, which is a
clone of Scala.js 1.x's test bridge, and diregard the native 0.6.x
implementation of the bridge. This ensures that our test adapter,
which is always from 1.x, can talk to the bridge.